pyproject.toml
[project]
name = "llm-lib-lag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "langchain>=0.3.17",
    "langchain-anthropic>=0.3.6",
    "langchain-community>=0.3.16",
    "langchain-fireworks>=0.2.7",
    "langchain-google-genai>=2.0.9",
    "langchain-groq>=0.2.4",
    "langchain-mistralai>=0.2.6",
    "langchain-openai>=0.3.3",
    "pydantic>=2.10.6",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]


[dependency-groups]
dev = ["ipykernel>=6.29.5", "ruff>=0.9.4", "pytest>=8.0.0"]

[tool.ruff]

[tool.ruff.lint]
select = [
    "N",    # Enable all pep8-naming rules
    "E7",   # Statements lints
    "F401", # Module imported but unused

    "UP006", # foo: list[int] = [1, 2, 3] instead of foo: List[int] = [1, 2, 3]
    "UP007", # Use X | Y for type annotations
    "UP017",

    "UP038", # Use isinstance(x, int | float) instead of isinstance(x, (int, float))
    "UP040",

    "UP045", # Use X | None instead of Optional[X]

    "ANN001", # Checks that function arguments have type annotations.
    "ANN201", # Missing return type annotation for public function {name}

    "SIM2", # Simplify boolean expressions

    # "T20", # Find uses of 'print' or 'pprint'

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
reportMatchNotExhaustive = true
typeCheckingMode = "strict"
reportUnknownParameterType = true
reportUntypedFunctionDecorator = true
reportMissingTypeStubs = true
reportUnusedImport = true
reportUnusedVariable = true
useLibraryCodeForTypes = true
